CORE-12218: More module moves
