+++ libs/rest/json-serialization/build.gradle
@@ -17,6 +17,7 @@ dependencies {
 
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
     implementation project(":libs:serialization:json-serializers")
+    implementation project(':libs:rest:rest')
 
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
 }
