+++ libs/rest/json-serialization/src/main/java/net/corda/rest/json/serialization/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.common.json.serialization;
+package net.corda.rest.json.serialization;
 
 import org.osgi.annotation.bundle.Export;
