+++ libs/rest/rest-client/src/main/kotlin/net/corda/rest/client/serialization/RestClientSerialization.kt
@@ -1,6 +1,5 @@
 package net.corda.rest.client.serialization
 
-import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParseException
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.BeanProperty
@@ -9,12 +8,12 @@ import com.fasterxml.jackson.databind.JavaType
 import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.deser.ContextualDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
-import net.corda.common.json.serialization.jacksonObjectMapper
+import net.corda.rest.json.serialization.jacksonObjectMapper
 import net.corda.crypto.core.parseSecureHash
 import net.corda.rest.JsonObject
 import net.corda.rest.durablestream.DurableCursorTransferObject
 import net.corda.rest.durablestream.api.Cursor
-import net.corda.rest.json.JsonObjectSerializer
+import net.corda.rest.json.serialization.JsonObjectSerializer
 import net.corda.utilities.trace
 import net.corda.v5.base.types.MemberX500Name
 import net.corda.v5.crypto.SecureHash
