+++ libs/rest/rest-common/build.gradle
@@ -11,8 +11,6 @@ dependencies {
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     implementation "net.corda:corda-base"
     implementation "net.corda:corda-application"
-    api "com.fasterxml.jackson.core:jackson-databind:$jacksonVersion"
-    implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
 
     api project(":libs:rest:rest")
     implementation project(':libs:messaging:messaging')
