--- libs/rest/rest-common/src/main/java/net/corda/rest/json/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.rest.json;
-
-import org.osgi.annotation.bundle.Export;
